Fixes issue where no apps were loaded
